app/requirements.txt
astroid==2.3.2
Django==2.2.6
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
djangorestframework==3.10.3
django-filter==2.2.0
markdown==3.1.1
requests==2.22.0
gunicorn==19.9.0
whitenoise==4.1.4
